refactor: relocate session-store to pi runner dir
DefaultSessionStore + SessionStoreManager are pi-SDK-coupled (wrap
SessionManager from @mariozechner/pi-coding-agent). Move them out of
src/legacy/core/ into src/agent/runners/pi/, alongside the runner that
actually owns this implementation. If we ever swap runners, this code
goes with it.

src/sessions/types.ts stays as the runner-agnostic interface — no
barrel re-export from sessions/, importers go directly to the pi path
to be honest about the coupling.

Two files merged into one (manager was 101 LOC, was a thin factory
over the store).

Layout aligned with openclaw, where src/sessions/ holds only
runner-agnostic concepts (id, lifecycle, label, policy) and
SessionManager usage lives next to the runner under src/agents/.
