+++ src/legacy/plugins/http/chat.test.ts
@@ -5,7 +5,7 @@ import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import http from "node:http";
 import { ApiServer } from "./server.js";
 import { CronScheduler } from "../../automation/cron-job.js";
-import { SessionStoreManager } from "../../core/session-store-manager.js";
+import { SessionStoreManager } from "../../../agent/runners/pi/session-store.js";
 import { AgentRuntime } from "../../../agent/runtime.js";
 import { createMockSessionStore } from "../../../test-helpers.js";
