+++ src/legacy/plugins/http/routes.ts
@@ -4,7 +4,7 @@ import type { ServerResponse } from "node:http";
 
 import type { CronScheduler } from "../../automation/cron-job.js";
 import type { ConfigReloader } from "../../workspace/config-reloader.js";
-import type { SessionStoreManager } from "../../core/session-store-manager.js";
+import type { SessionStoreManager } from "../../../agent/runners/pi/session-store.js";
 import type { HookRegistry } from "../../plugins/hooks.js";
 import type { AgentRuntime } from "../../../agent/runtime.js";
 import type { ApiRequest } from "./middleware.js";
