+++ src/legacy/plugins/types.ts
@@ -5,7 +5,7 @@ import type { Logger } from "../../logging/logger.js";
 import type { AgentMessage, AgentTool } from "@mariozechner/pi-agent-core";
 import type { SessionStore } from "../../sessions/types.js";
 import type { Transport } from "../../gateway/types.js";
-import type { SessionStoreManager } from "../core/session-store-manager.js";
+import type { SessionStoreManager } from "../../agent/runners/pi/session-store.js";
 import type { IsotopesConfigFile } from "../../config.js";
 import type { LazyTransportContext } from "../tools/react.js";
 import type { AgentRuntime } from "../../agent/runtime.js";
